# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  isPremium: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  isPremium: boolean;
  createdAt: string;
  totalChats: number;
  totalMessages: number;
}

export interface UserProfileUpdate {
  /** @minLength 3 */
  username?: string;
}

export interface Chat {
  id: number;
  title: string;
  userId: number;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
}

export interface ChatInput {
  /** @minLength 1 */
  title: string;
}

export interface ChatUpdate {
  /** @minLength 1 */
  title?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  chatId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export interface ModelStatus {
  name: string;
  version: string;
  isReady: boolean;
  vocabSize: number;
  totalTrainingSamples: number;
  /** @nullable */
  lastTrainedAt: string | null;
}

export interface TrainResult {
  success: boolean;
  samplesUsed: number;
  message: string;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  currency: string;
  period: string;
  features: string[];
  isPopular: boolean;
  /** @nullable */
  description?: string | null;
}