# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QuantumAI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: chats
    description: Chat management
  - name: messages
    description: Message operations
  - name: model
    description: AI model operations
  - name: users
    description: User operations
  - name: plans
    description: Premium plan operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List all chats for current user
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a single chat by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update a chat title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatUpdate"
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      operationId: deleteChat
      tags: [chats]
      summary: Delete a chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /chats/{chatId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages for a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePair"

  /model/status:
    get:
      operationId: getModelStatus
      tags: [model]
      summary: Get the quantum-ai-0.1 model status
      responses:
        "200":
          description: Model status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelStatus"

  /model/train:
    post:
      operationId: trainModel
      tags: [model]
      summary: Trigger model training on recent chat history
      responses:
        "200":
          description: Training result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainResult"

  /users/me:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all premium plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    User:
      type: object
      required: [id, username, email, isPremium, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time

    UserProfile:
      type: object
      required: [id, username, email, isPremium, createdAt, totalChats, totalMessages]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        isPremium:
          type: boolean
        createdAt:
          type: string
          format: date-time
        totalChats:
          type: integer
        totalMessages:
          type: integer

    UserProfileUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3

    Chat:
      type: object
      required: [id, title, userId, createdAt, updatedAt, messageCount]
      properties:
        id:
          type: integer
        title:
          type: string
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        lastMessage:
          type: ["string", "null"]

    ChatInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1

    ChatUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1

    Message:
      type: object
      required: [id, chatId, role, content, createdAt]
      properties:
        id:
          type: integer
        chatId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    MessagePair:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: "#/components/schemas/Message"
        assistantMessage:
          $ref: "#/components/schemas/Message"

    ModelStatus:
      type: object
      required: [name, version, isReady, vocabSize, totalTrainingSamples, lastTrainedAt]
      properties:
        name:
          type: string
        version:
          type: string
        isReady:
          type: boolean
        vocabSize:
          type: integer
        totalTrainingSamples:
          type: integer
        lastTrainedAt:
          type: ["string", "null"]

    TrainResult:
      type: object
      required: [success, samplesUsed, message]
      properties:
        success:
          type: boolean
        samplesUsed:
          type: integer
        message:
          type: string

    Plan:
      type: object
      required: [id, name, price, currency, period, features, isPopular]
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        currency:
          type: string
        period:
          type: string
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
        description:
          type: ["string", "null"]