# lib/api-zod/src/generated/types/chat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Chat {
  id: number;
  title: string;
  userId: number;
  createdAt: Date;
  updatedAt: Date;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
}