# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  /** @minLength 1 */
  title: string;
}