# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  /** @minLength 3 */
  username?: string;
}